split user shared up better
